backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
bleach==6.2.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
cssselect2==0.7.0
Django==5.1.6
django-allauth==65.5.0
django-cors-headers==4.5.0
django-markdownify==0.9.5
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
fonttools==4.56.0
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
html5lib==1.1
httplib2==0.22.0
idna==3.10
Markdown==3.7
markdownify==1.1.0
multidict==6.1.0
openai==0.28.0
pillow==11.1.0
propcache==0.2.1
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyparsing==3.2.3
pyphen==0.17.2
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9.1
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tinycss2==1.4.0
tqdm==4.67.1
typing-extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
weasyprint==57.1
webencodings==0.5.1
yarl==1.18.3
zopfli==0.2.3.post1
firebase-admin==6.6.0